refactor(Application) : chess 패키지 안으로 이동
