+++ src/main/java/chess/SparkChessApplication.java
@@ -1,4 +1,4 @@
-package Application;
+package chess;
 
 import spark.ModelAndView;
 import spark.template.handlebars.HandlebarsTemplateEngine;
