+++ src/main/java/chess/SpringChessApplication.java
@@ -1,4 +1,4 @@
-package Application;
+package chess;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
